updating types to export out
