+++ src/components/Avatar/Avatar.module.scss
@@ -6,8 +6,6 @@
   align-items: center;
   justify-content: center;
   flex-shrink: 0;
-  width: 50px;
-  height: 50px;
   border-radius: 50%;
   overflow: hidden;
   user-select: none;
