+++ src/components/Avatar/Avatar.tsx
@@ -1,19 +1,19 @@
 import React from 'react';
 import styles from './Avatar.module.scss';
- 
+
 export type AvatarPropsT = {
   src: string;
   alt: string;
+  size: number;
   classProp?: string;
 };
 
-const Avatar = ({
-  src,
-  alt,
-  classProp = '',
-}: AvatarPropsT) => {
+const Avatar = ({ src, alt, size = 50, classProp = '' }: AvatarPropsT) => {
   return (
-    <div className={`${styles.avatar} ${classProp}`}>
+    <div
+      className={`${styles.avatar} ${classProp}`}
+      style={{ height:`${size}px`, width:`${size}px`}}
+    >
       <img src={src} alt={alt} />
     </div>
   );
