+++ src/components/Button/Button.tsx
@@ -1,8 +1,27 @@
 import React, { MouseEventHandler } from 'react';
-import { ButtonT, ButtonCategoriesE, ButtonSizesE } from '../../types/Form';
-import { IconT } from '../../types/General';
 import styles from './Button.module.scss';
-import Icon from '../Icon/Icon';
+import Icon, { IconT } from '../Icon/Icon';
+
+export type ButtonT = 'button' | 'submit' | 'reset';
+
+/**
+ * Enums are formatted in snake_case so that we can use the mapped
+ * value as a SCSS value.
+ */
+export enum ButtonCategoriesE {
+  PRIMARY_SOLID = 'primary_solid',
+  PRIMARY_OUTLINE = 'primary_outline',
+  PRIMARY_CLEAR = 'primary_clear',
+  SECONDARY_SOLID = 'secondary_solid',
+  SECONDARY_OUTLINE = 'secondary_outline',
+  SECONDARY_CLEAR = 'secondary_clear',
+}
+
+export enum ButtonSizesE {
+  SMALL = 'small',
+  MEDIUM = 'medium',
+  LARGE = 'large',
+}
 
 export type ButtonPropsT = {
   active?: boolean;
